refactor(wip): prettier

+++ contracts/contracts/CalculatedPriceProvider.sol
@@ -9,8 +9,12 @@ import {CalculatedPriceProviderStorage} from "./CalculatedPriceProviderStorage.s
 
 /// @title This contract gets the price from some IPriceProviders and do the math to calculate
 /// a deduced price, for example RIFBTC and BTCUSD gives the price of RIFUSD
-contract CalculatedPriceProvider is CalculatedPriceProviderStorage, IPriceProvider, IPriceProviderRegisterEntry {
-    using SafeMath for uint;
+contract CalculatedPriceProvider is
+    CalculatedPriceProviderStorage,
+    IPriceProvider,
+    IPriceProviderRegisterEntry
+{
+    using SafeMath for uint256;
 
     /**
     Contract creation.
@@ -22,16 +26,20 @@ contract CalculatedPriceProvider is CalculatedPriceProviderStorage, IPriceProvid
     @param _divisor base value used to scale the result by dividing it.
     @param _divideBy list of IPriceProvider to query the and then divide the result
     */
-    function initialize(IGovernor _governor, address[] calldata _wlist,
-        uint _multiplicator, IPriceProvider[] calldata _multiplyBy,
-        uint _divisor, IPriceProvider[] calldata _divideBy) external initializer {
+    function initialize(
+        IGovernor _governor,
+        address[] calldata _wlist,
+        uint256 _multiplicator,
+        IPriceProvider[] calldata _multiplyBy,
+        uint256 _divisor,
+        IPriceProvider[] calldata _divideBy
+    ) external initializer {
         _initialize(_governor);
         calculatedPriceProviderData._initialize(_multiplicator, _multiplyBy, _divisor, _divideBy);
         for (uint256 i = 0; i < _wlist.length; i++) {
             iterableWhitelistData._addToWhitelist(_wlist[i]);
         }
     }
-
 
     /**
      * @dev Add to the list of contracts that can stake in this contract
@@ -46,7 +54,10 @@ contract CalculatedPriceProvider is CalculatedPriceProviderStorage, IPriceProvid
      * @param _whitelisted - the override coinPair
      * @param _hint Optional hint to start traversing the array, zero is to search all the array.
      */
-    function removeFromWhitelist(address _whitelisted, uint256 _hint) external onlyAuthorizedChanger() {
+    function removeFromWhitelist(address _whitelisted, uint256 _hint)
+        external
+        onlyAuthorizedChanger()
+    {
         iterableWhitelistData._removeFromWhitelist(_whitelisted, _hint);
     }
 
@@ -70,7 +81,13 @@ contract CalculatedPriceProvider is CalculatedPriceProviderStorage, IPriceProvid
     Get the calculated price
 
     */
-    function peek() external override view whitelistedOrExternal(iterableWhitelistData) returns (bytes32, bool) {
+    function peek()
+        external
+        override
+        view
+        whitelistedOrExternal(iterableWhitelistData)
+        returns (bytes32, bool)
+    {
         (uint256 price, bool valid) = calculatedPriceProviderData._peek();
         return (bytes32(price), valid);
     }
